Added PutShipManager
